+++ poetry/console/commands/__init__.py
@@ -5,6 +5,7 @@ from .config import ConfigCommand
 from .install import InstallCommand
 from .lock import LockCommand
 from .new import NewCommand
+from .publish import PublishCommand
 from .remove import RemoveCommand
 from .show import ShowCommand
 from .update import UpdateCommand